Refactored remaining routes and forms into main blueprint
