+++ app/main/routes.py
@@ -1,38 +1,38 @@
 # -*- coding: utf-8 -*-
 """
-    routes
-    ~~~~~~
+    main.routes
+    ~~~~~~~~~~~
 
-    Routes for the different parts of the data ecosystem.
+    Routes that do not fit with any of the more specific blueprints or are
+    core features of the application.
 """
 from flask import abort, render_template, send_from_directory
 from flask_login import current_user, login_required
 
 from app import app
+from app.main import bp
 
 
-@app.route('/')
+@bp.route('/')
 def index():
     return render_template('index.html', title='Home Page')
 
 
-@app.route('/error/test_500')
+@bp.route('/error/test_500')
 def error_test_500():
     abort(500)
 
 
-@app.route('/profile')
+@bp.route('/profile')
 @login_required
 def profile():
     user_data = 'abc'
     return render_template('profile.html', title='Profile', data=user_data)
 
 
-@app.route('/uploads/<filename>')
+@bp.route('/uploads/<filename>')
 @login_required
 def uploaded_file(filename):
     if current_user.role.role != 'admin':
         abort(403)
     return send_from_directory(app.config['IMPORT_FOLDER'], filename)
-
-
