+++ app/templates/base.html
@@ -20,7 +20,7 @@
         <ul class="navbar-nav">
             {% if not current_user.is_anonymous %}
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('profile') }}">Profile</a>
+                    <a class="nav-link" href="{{ url_for('main.profile') }}">Profile</a>
                 </li>
                 {% if current_user.role.role == 'admin' %}
                     <li class="nav-item">
