+++ app/templates/errors/500.html
@@ -3,5 +3,5 @@
 {% block app_content %}
     <h1>An unexpected error has occurred</h1>
     <p>The administrator has been notified.</p>
-    <p><a href="{{ url_for('index') }}">Back</a></p>
+    <p><a href="{{ url_for('main.index') }}">Back</a></p>
 {% endblock %}
